requirements.txt
beautifulsoup4==4.5.1
Django==1.10
djangorestframework==3.5.2
docutils==0.12
elasticsearch==2.3.0
konlpy==0.4.4
lxml==3.6.4
mysqlclient==1.3.7
nltk==3.2.1
urllib3==1.19
JPype1-py3==0.5.5.2